03-3-Langgraph/03-实战案例/01-搜索相关主题回答问题/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
pydantic-settings==2.0.3
motor==3.3.1
langgraph==0.0.36
langchain==0.0.314
langchain-community==0.0.13
bs4==0.0.1
python-dotenv==1.0.0
aiohttp==3.8.5
chromadb==0.4.17
dashscope==1.13.3 